Apply Dan suggestion to group interface with type guard
