+++ package.json
@@ -30,6 +30,14 @@
     "morgan": "^1.10.0"
   },
   "devDependencies": {
+    "@types/body-parser": "^1.19.0",
+    "@types/express": "^4.17.7",
+    "@types/express-serve-static-core": "^4.17.8",
+    "@types/http-errors": "^1.6.3",
+    "@types/jest": "^26.0.4",
+    "@types/morgan": "^1.9.1",
+    "@typescript-eslint/eslint-plugin": "^3.6.0",
+    "@typescript-eslint/parser": "^3.6.0",
     "eslint": "^7.4.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
@@ -38,14 +46,6 @@
     "prettier": "^2.0.5",
     "ts-jest": "^26.1.1",
     "tsc-watch": "^4.2.9",
-    "typescript": "^3.9.6",
-    "@types/body-parser": "^1.19.0",
-    "@types/express": "^4.17.7",
-    "@types/express-serve-static-core": "^4.17.8",
-    "@types/http-errors": "^1.6.3",
-    "@types/jest": "^26.0.4",
-    "@types/morgan": "^1.9.1",
-    "@typescript-eslint/eslint-plugin": "^3.6.0",
-    "@typescript-eslint/parser": "^3.6.0"
+    "typescript": "^3.9.6"
   }
 }
