--- src/is/index.ts
@@ -1,2 +1,0 @@
-export * from './basicLegacyError';
-export * from './legacyError';
