+++ src/types/errors/LegacyError.ts
@@ -1,4 +1,11 @@
-import { ILegacyError } from '../types/error_types';
+import { IBasicLegacyError } from './BasicLegacyError';
+
+/**
+ * Format of some older error messages.
+ */
+export interface ILegacyError extends IBasicLegacyError {
+	statusCode: number;
+}
 
 /**
  * Type guard to check if something is a subset of the interface LegacyError.
