+++ src/types/errors/index.ts
@@ -1,0 +1,3 @@
+export * from './BasicLegacyError';
+export * from './LegacyError';
+export * from './TxLegacyError';
